Fixed faulty overlap of drawings
Fixed issue where, when painting, your drawings would be shown behind the previous part (top or middle parts drawn by other player), but in the end screen it would be the other way around (your part above the previous part). This is fixed now so that you can see that you're drawing on top of the previous part.